+++ src/io/segment/android/models/Screen.java
@@ -1,0 +1,37 @@
+package io.segment.android.models;
+
+import java.util.Calendar;
+
+import org.json.JSONObject;
+
+public class Screen extends Track {
+	
+	public final static String ACTION = "screen";
+	
+	private static final String SCREEN_KEY = "screen";
+	
+	public Screen (JSONObject obj) {
+		super(obj);
+	}
+	
+	public Screen(String userId, 
+				 String screen, 
+				 EventProperties properties, 
+				 Calendar timestamp,
+				 Context context) {
+
+		super(userId, "Viewed " + screen, properties, timestamp, context);
+
+		put("action", ACTION);
+		
+		setScreen(screen);
+	}
+
+	public String getScreen() {
+		return this.optString(SCREEN_KEY, null);
+	}
+
+	public void setScreen(String screen) {
+		this.put(SCREEN_KEY, screen);
+	}
+}