+++ src/io/segment/android/providers/AmplitudeProvider.java
@@ -4,6 +4,7 @@ import io.segment.android.errors.InvalidSettingsException;
 import io.segment.android.models.EasyJSONObject;
 import io.segment.android.models.EventProperties;
 import io.segment.android.models.Identify;
+import io.segment.android.models.Screen;
 import io.segment.android.models.Track;
 import io.segment.android.models.Traits;
 import io.segment.android.provider.SimpleProvider;
@@ -62,6 +63,12 @@ public class AmplitudeProvider extends SimpleProvider {
 		
 		Amplitude.setUserId(userId);
 		Amplitude.setGlobalUserProperties(traits);
+	}
+
+	@Override
+	public void screen(Screen screen) {
+		// track a "Viewed SCREEN" event
+		track(screen);
 	}
 	
 	@Override