+++ src/io/segment/android/providers/CrittercismProvider.java
@@ -3,6 +3,7 @@ package io.segment.android.providers;
 import io.segment.android.errors.InvalidSettingsException;
 import io.segment.android.models.EasyJSONObject;
 import io.segment.android.models.Identify;
+import io.segment.android.models.Screen;
 import io.segment.android.models.Track;
 import io.segment.android.models.Traits;
 import io.segment.android.provider.SimpleProvider;
@@ -83,6 +84,12 @@ public class CrittercismProvider extends SimpleProvider {
 			
 			Crittercism.setMetadata(traits);
 		}
+	}
+	
+	@Override
+	public void screen(Screen screen) {
+		// track a "Viewed SCREEN" event
+		track(screen);
 	}
 	
 	@Override