+++ src/io/segment/android/providers/FlurryProvider.java
@@ -5,6 +5,7 @@ import io.segment.android.errors.InvalidSettingsException;
 import io.segment.android.models.EasyJSONObject;
 import io.segment.android.models.EventProperties;
 import io.segment.android.models.Identify;
+import io.segment.android.models.Screen;
 import io.segment.android.models.Track;
 import io.segment.android.models.Traits;
 import io.segment.android.provider.SimpleProvider;
@@ -103,6 +104,15 @@ public class FlurryProvider extends SimpleProvider {
 
 		String userId = identify.getUserId();
 		FlurryAgent.setUserId(userId);
+	}
+	
+
+	@Override
+	public void screen(Screen screen) {
+		// increment flurry's page view count
+		FlurryAgent.onPageView();
+		// track a "Viewed SCREEN" event
+		track(screen);
 	}
 	
 	@Override